requirements.txt
openai>=1.30.0
httpx>=0.27.0
pyyaml>=6.0
rich>=13.7.0
click>=8.1.0
fastapi>=0.115.0
uvicorn>=0.30.0
opencv-python>=4.9.0
numpy>=1.24.0
